+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - fix Colour display
+- fix 16-bit alpha scaling
 
 ## 9.0.0-10 29025/02/18